+++ views/account/login.php
@@ -8,15 +8,15 @@ echo $model->display_messages();
 <form action="<?php echo $model->getUrl( 'account/login' ); ?>" method="post">
     <ul>
         <li>
-            <label>Email</label>
+            <label><?php echo _( 'Email' ); ?></label>
             <input type="email" name="email" required value="<?php echo $model->getFieldValue( 'email' ); ?>" />
         </li>
         <li>
-            <label>Password</label>
+            <label><?php echo _( 'Password' ); ?></label>
             <input type="password" name="password" required value="<?php echo $model->getFieldValue( 'password' ); ?>" />
         </li>
         <li>
-            <button type="submit">Login</button>
+            <button type="submit"><?php echo _( 'Login' ); ?></button>
         </li>
     </ul>
 </form>