+++ views/install.php
@@ -6,21 +6,21 @@ echo $this->getModule( 'before-header' );
 <?php echo $model->display_messages(); ?>
 <form action="" autocomplete="off" method="post">
     <fieldset class="cssc-form">
-        <div class="box"><label for="bs-name">Site Name :</label><input required name="bs-name" id="bs-name" type="text" value="<?php echo $model->getFieldValue( 'bs-name' ); ?>" /></div>
+        <div class="box"><label for="bs-name"><?php echo _( 'Site Name:' ); ?></label><input required name="bs-name" id="bs-name" type="text" value="<?php echo $model->getFieldValue( 'bs-name' ); ?>" /></div>
     </fieldset>
     <fieldset class="cssc-form">
-        <div class="box"><label for="bs-adminemail">Admin email :</label><input required name="bs-adminemail" id="bs-adminemail" type="email" value="<?php echo $model->getFieldValue( 'bs-adminemail' ); ?>" /></div>
-        <div class="box"><label for="bs-adminpass">Pass :</label><input required name="bs-adminpass" id="bs-adminpass" type="password" value="<?php echo $model->getFieldValue( 'bs-adminpass' ); ?>" /></div>
+        <div class="box"><label for="bs-adminemail"><?php echo _( 'Admin email:' ); ?></label><input required name="bs-adminemail" id="bs-adminemail" type="email" value="<?php echo $model->getFieldValue( 'bs-adminemail' ); ?>" /></div>
+        <div class="box"><label for="bs-adminpass"><?php echo _( 'Pass:' ); ?></label><input required name="bs-adminpass" id="bs-adminpass" type="password" value="<?php echo $model->getFieldValue( 'bs-adminpass' ); ?>" /></div>
     </fieldset>
     <fieldset class="cssc-form">
-        <div class="box"><label for="bs-dbhost">Host :</label><input required name="bs-dbhost" id="bs-dbhost" type="text" value="<?php echo $model->getFieldValue( 'bs-dbhost' ); ?>" /></div>
-        <div class="box"><label for="bs-dbname">Name :</label><input required name="bs-dbname" id="bs-dbname" type="text" value="<?php echo $model->getFieldValue( 'bs-dbname' ); ?>" /></div>
-        <div class="box"><label for="bs-dbuser">User :</label><input required name="bs-dbuser" id="bs-dbuser" type="text" value="<?php echo $model->getFieldValue( 'bs-dbuser' ); ?>" /></div>
-        <div class="box"><label for="bs-dbpass">Pass :</label><input required name="bs-dbpass" id="bs-dbpass" type="text" value="<?php echo $model->getFieldValue( 'bs-dbpass' ); ?>" /></div>
-        <div class="box"><label for="bs-dbprefix">Prefix :</label><input required name="bs-dbprefix" id="bs-dbprefix" type="text" value="<?php echo $model->getFieldValue( 'bs-dbprefix' ); ?>" /></div>
+        <div class="box"><label for="bs-dbhost"><?php echo _( 'Host:' ); ?></label><input required name="bs-dbhost" id="bs-dbhost" type="text" value="<?php echo $model->getFieldValue( 'bs-dbhost' ); ?>" /></div>
+        <div class="box"><label for="bs-dbname"><?php echo _( 'Name:' ); ?></label><input required name="bs-dbname" id="bs-dbname" type="text" value="<?php echo $model->getFieldValue( 'bs-dbname' ); ?>" /></div>
+        <div class="box"><label for="bs-dbuser"><?php echo _( 'User:' ); ?></label><input required name="bs-dbuser" id="bs-dbuser" type="text" value="<?php echo $model->getFieldValue( 'bs-dbuser' ); ?>" /></div>
+        <div class="box"><label for="bs-dbpass"><?php echo _( 'Pass:' ); ?></label><input required name="bs-dbpass" id="bs-dbpass" type="text" value="<?php echo $model->getFieldValue( 'bs-dbpass' ); ?>" /></div>
+        <div class="box"><label for="bs-dbprefix"><?php echo _( 'Prefix:' ); ?></label><input required name="bs-dbprefix" id="bs-dbprefix" type="text" value="<?php echo $model->getFieldValue( 'bs-dbprefix' ); ?>" /></div>
     </fieldset>
     <div>
-        <button class="cssc-button" type="submit">Submit</button>
+        <button class="cssc-button" type="submit"><?php echo _( 'Submit' ); ?></button>
     </div>
 </form>
 <?php