+++ views/modules/header.php
@@ -5,7 +5,7 @@ $model = $this->getModel();
 <ul>
     <li><a href="<?php echo $model->getUrl( ); ?>"><?php echo _( 'Home' ); ?></a></li>
     <?php if ( $user->isLoggedIn() ) { ?>
-    <li><a href="<?php echo $model->getUrl( 'account/dashboard' ); ?>"><?php echo $user->getEmail(); ?></a></li>
+    <li><img src="<?php echo $user->getAvatar(12); ?>" alt="" /> <a href="<?php echo $model->getUrl( 'account/dashboard' ); ?>"><?php echo $user->getName(); ?></a></li>
     <li><a href="<?php echo $model->getUrl( 'account/logout' ); ?>"><?php echo _( 'Logout' ); ?></a></li>
     <?php } else { ?>
     <li><a href="<?php echo $model->getUrl( 'account/login' ); ?>"><?php echo _( 'Login' ); ?></a></li>